Move gnomAD subsetting, annotation & MAPS score functions to separate modules
