+++ scripts/maps_gcp.py
@@ -16,10 +16,10 @@ See rules/evaluation.smk for a generic hailctl command.
 """
 
 import hail as hl
-import gnomad.utils.vep
 
-from prepare_gnomad import get_worst_consequence_with_non_coding, prepare_ht
-from maps_score import maps
+from annotate_gnomad import filter_gnomad, annotate_for_maps  # pylint: disable=E0401
+from maps import maps  # pylint: disable=E0401
+
 
 if __name__ == '__main__':
     import argparse
@@ -58,6 +58,7 @@ if __name__ == '__main__':
 
     hl.init(default_reference=args.genome_assembly)
 
+    # TODO: include annotation
     # intervals = hl.import_bed(bed_file, reference_genome=reference_genome')
     with open(args.intervals, 'r') as f:
         intervals = f.readlines()
@@ -67,27 +68,15 @@ if __name__ == '__main__':
     ]
 
     mutation_ht = hl.read_table(args.mutation_ht)
-
-    # subset context table
     context_ht = hl.read_table(args.context_ht)
-    context_ht = hl.filter_intervals(context_ht, intervals)
 
-    # prepare gnomAD table
     ht = hl.read_table(args.gnomAD_ht)
-    ht = hl.filter_intervals(ht, intervals)
-    ht = ht.filter(hl.len(ht.filters) == 0)
-    ht = gnomad.utils.vep.filter_vep_to_canonical_transcripts(ht)
-    print(f'entries in gnomAD after filtering: {ht.count()}')
-
-    ht = get_worst_consequence_with_non_coding(ht)
-    context = context_ht[ht.key]
-    snp_ht = prepare_ht(
-        ht=ht.annotate(context=context.context, methylation=context.methylation),
-        trimer=True,
-    )
+    ht = filter_gnomad(ht, intervals)
+    ht = annotate_for_maps(ht, context_ht)
 
     print('MAPS score')
-    maps_ht = maps(snp_ht, mutation_ht, additional_grouping=['protein_coding'])
+    # TODO: use different grouping
+    maps_ht = maps(ht, mutation_ht, additional_grouping=['protein_coding'])
     maps_ht.show()
 
     print('save...')
